+++ hopper-xtreme/common/src/main/java/com/chimericdream/hopperxtreme/block/HopperVariantBlock.java
@@ -10,4 +10,6 @@ public interface HopperVariantBlock {
     boolean isWithFilter();
 
     int getCooldownInTicks();
+
+    int getItemsPerTick();
 }